Improve synchronization efficiency.
